% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codificacion.R
\name{codificacion}
\alias{codificacion}
\title{Codificacion}
\usage{
codificacion(x, desde = "UTF-8", hacia = "ISO8859-1")
}
\arguments{
\item{x}{Dataframe a convertir.}

\item{desde}{Codificacion origen.}

\item{hacie}{Codificacion destino.}
}
\description{
Codifica un dataframe de una codificacion origen a otra.
Por defecto convierte de UTF-8 a ISO8859-1, o mas conocido como LATIN1, pero se puede especificar otra.
}
\examples{
diremail("D:/Ejemplos")
x <- leer(c("archivo","carpeta$"))
x <- codificacion(x)
-En este punto, si el archivo de origen tiene la codifiacion UTF-8, se convertira a LATIN1, de a cuerdo al alfabeto latino.
}
