% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diremail.R
\name{diremail}
\alias{diremail}
\title{Especificacion de la carpeta general y el correo de google}
\usage{
diremail(c = "", email = "")
}
\description{
Especificacion de la carpeta general y el correo de google
}
