% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_sd.R
\name{drive_sd}
\alias{drive_sd}
\title{Descarga y subida de archivos en Google drive}
\usage{
drive_sd(mod = NULL, arch = c("nom", "nom2", "dir"), ...)
}
\description{
Sube o descarga archivos de google drive
}
