% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escribir.R
\name{escribir}
\alias{escribir}
\title{Escritura simplificada (Incompleto)}
\usage{
escribir(base, arch = c("", ""), sep = "\\t", ...)
}
\arguments{
\item{base}{DataFrame a escribir.}

\item{arch}{Vector donde se indica Nombre del archivo y extension, para arch[1] y patron de la carpeta en donde se va a escribir en arch[2].}

\item{sep}{Por defecto, si se imprime un archivo .txt este se separara con tabulaciones, se puede editar para usar cualquier separador.}

\item{...}{Parametros propios de las funciones  write.csv() y  write.table().}
}
\description{
Escribe un archvo .csv o .txt por el momento, tal como la funcion leer() se indica el nombre y el patron a usar en la carpeta indicada en diremail().
}
\examples{
Hola
}
