% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer.R
\encoding{LATIN1}
\name{leer}
\alias{leer}
\title{Lectura de multiples fuentes}
\usage{
leer(arch = c("", ""), delim = NULL, ...)
}
\arguments{
\item{arch}{Direccion completa o vector donde indica el nombre y la ruta en forma de patron, si se lee un gsheet se indica "gsheet" en arch[2],
si se requiere leer por su ID se debe especificar en el mismo "gsheet.ID", para esto es necesario especificar el correo electronico a usar para leer los archivos de tipo gsheet,
indicado en diremail()}

\item{delim}{Opcional. si se lee un archivo .txt en ella se indica el delimitador a usar.}

\item{...}{Parametros propios de las distintas funciones read_csv(), read_delim(), read_excel() y range_speedread().}
}
\value{
Un data frame del archivo leido
}
\description{
Lee archivos de varios tipos (.xlsx, .csv, .txt y gsheets) , no es necesario especificar la extension ya que este la detecta de forma automatica.
Puede leer una direccion completa o un patron de una lista de direcciones indicada en diremail().
Por defecto, todos los archivos a leer convierte las columnas a tipo "character", puede haber problemas con fechas en archivos .xlsx para columnas de tipo "Date"
debido al mismo motivo, se toma en cuenta para solucionarlo.
}
\examples{
-Lectura simple de una ruta completa.

leer("D:/archivo.csv")

-En este caso se toma como ejemplo un archivo cuya direccion tiene como patron terminar
-en la palabra "Carpeta", la direccion completa puede ser "D:/Documentos/Archivos/Carpetas/Carpeta2/archivo_2.xlsx",
-no es necesario escribir la extension, se puede observar como se puede usar un parametro propio de otra funcion,
-en este caso de read_csv() el parametro col_select

leer(c("archivo_2","Carpeta2$"),col_select = c("Matricula","Programa"))

-Para leer un gsheet puede ser por su nombre o ID, se pueden usar los parametros propios de
-la funcion range_speedread().

leer(c("Info General","gsheet"))
leer(c("sDfGtHjGfusj56G8hkohejemplo","gsheet.ID"), sheet= "Info")
}
