% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repartir.R
\name{repartir}
\alias{repartir}
\title{Reparte equitativamente}
\usage{
repartir(Base, col, x, reem = NA)
}
\arguments{
\item{Base}{Dataframe, se retorna el mismo con la columna repartida}

\item{col}{Columna del dataframe que se tomara para repartir el vector.}

\item{x}{Vector de datos los cuales se van a repartir.}

\item{reem}{Datos que se van a reemplazar de la columna dada, por defecto es "NA".}
}
\description{
La funcion reparte de una manera lo mas equitativa posible un vector, se puede elegir la columna y que datos se van a repartir y sustituir.
No incluye una opcion para repartir una columna completa ya que se puede utilizar directamente rep_len() para esas situaciones.
Los datos de reemlazo y los datos a reemplazar deben ser del mismo tipo.
}
\examples{
Base <- repartir(Base,c("A","B","C"),"columna")
Base <- repartir(Base,c(1,2,3)"columna2",c(3,4,5))
Base <- repartir(Base,c("A","B","C")"columna2",c("A","C","D"))
}
