% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unificar.R
\name{unificar}
\alias{unificar}
\title{Unificacion de varios archivos a la vez}
\usage{
unificar(
  dir = c(.c, ""),
  renom = NULL,
  cols = NULL,
  col_vac = "_",
  iden = "no",
  iden_esp = NULL,
  ...
)
}
\arguments{
\item{dir}{Ruta de la carpeta en donde contiene todos los archivos a unificar, puede utilizar un patron especificando un vector c("dir","pat")
para especificar archivos exactos.}

\item{renom}{Renombra las columnas al nombre que se necesiten}

\item{cols}{Vector de columnas a seleccionar en el DataFrame final, es muy util en archivos de tipo .xlsx ya que no contiene un parametro especifico
para seleccionar columnas, tambien es util cuando se requiere cambiar primero los nombres que no coinciden en todos los archivos a unificar, si una
columna dentro del archivo no existe, crea una columna con un solo caracter especificado en col_vac}

\item{col_vac}{Caracter utilizado para especificar como se debe llenar la columna vacia, en caso de que no exista en el archivo cuando se selecciona}

\item{...}{Parametros propios de las funciones que utiliza leer()}
}
\value{
Un DataFrame con todos los archivos unificados.
}
\description{
Unifica archivos completos en un solo DataFrame, utiliza la misma funcion leer() para transformar los archivos.
}
\examples{
-Aqui no se especifica ningun parametro mas que dir y pat, en donde unifica todos los archivos
-que contengan el patron "Actas Docente"

unificar("SIR$","Actas Docente")

-En este caso complejo se observa que no se selecciona ningun patron, toma todos los archivos
-que contenga la carpeta "FLOKZU", internamente primero lee el archivo, despues renombra
-las columnas que se requieran especificadas en "renom" si es que existen, despues selecciona
-las columnas especificadas en "cols", si no existe alguna columna, la crea por cada archivo
-y ya que no se especifica "col_vac", se llenara la columna con un "_", y por ultimo se
-especifica en un parametro propio de "read_excel()" las filas a omitir ya que todos
-los archivos son ".xlsx", se omite 1 fila.

Base <- unificar("FLOKZU$",
cols = c("Identificador","Matriicula","Correo","Fecha Inicio","Finalizado","Fecha Finalizacioon","Tarea Actual","Fecha Tarea","Analista SER"),
renom = c("SIU"="Matriicula",
          "Matricula"="Matriicula",
          "Correo del estudiante"="Correo",
          "Correo del prospecto"="Correo",
          "Fecha asignacioon Tarea actual"="Fecha Tarea",
          "A.analista SER"="Analista SER"),
          skip = 1)
}
