% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diremail.R
\name{diremail}
\alias{diremail}
\title{Especificacion de la carpeta general y el correo de google}
\usage{
diremail(c = "", email = "")
}
\arguments{
\item{c}{Carpeta origen.}

\item{email}{Correo de google para la carpeta del drive.}
}
\description{
Especifica la carpeta origen a utilizar para todos los archivos que se van a leer y tambien el correo a utilizar
para la carpeta del drive personal.
}
\examples{
diremail("D:/Reportes","ejemplo.1@gmail.com")
}
