% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diremail.R
\name{diremail}
\alias{diremail}
\title{Especificacion de la carpeta general, correo de google y conexion a MySQL}
\usage{
diremail(c = "", email = "", msql = "")
}
\arguments{
\item{c}{Carpeta origen.}

\item{email}{Correo de google para la carpeta del drive.}

\item{msql}{Conexion a MySQL, vector con orden c("host","usuario","contrasenia","base de datos")}
}
\description{
Especifica la carpeta origen a utilizar para todos los archivos que se van a leer, el correo a utilizar
para la carpeta del drive personal y especificaciones para MySQL
}
\examples{
diremail("D:/Reportes","ejemplo.1@gmail.com",c("10.10.10.100","usuario","contra","base_de_datos"))
}
